refactor: consolidate common utility classes for Flink CDC read (#18436)
* refactor: consolidate common utility classes for Flink CDC read
* resolve shuo's comments
